Function symphonia::default::get_probe

pub fn get_probe() -> &'static Probe

Gets the default Probe. This registry pre-registers all the formats selected by the feature flags in the includer’s Cargo.toml. If features is not set, the default set of Symphonia formats is registered.

This function is lazy and does not instantiate the Probe until the first call to this function.